requirements.txt
flask
flask-cors
python-dotenv
pymongo
werkzeug
langchain
langchain-huggingface
langchain-ollama
pinecone-client
sentence-transformers
PyPDF2
pypdf